OT275-93: renamed package common to builder
