+++ src/test/java/com/alkemy/ong/builder/UpdateOrganizationRequestBuilder.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.common;
+package com.alkemy.ong.builder;
 
 import com.alkemy.ong.infrastructure.rest.request.UpdateOrganizationRequest;
